Split operators.go into unary.go and binary.go
